+++ src/R/cytoprofiling/R/export.R
@@ -5,7 +5,7 @@ library(Seurat)
 #' @param df Cell table (see load_cytoprofiling)
 #' @param panel_json JSON data for Panel.json (see load_panel)
 #' @param use_gene_names If TRUE, use gene names as rownames, otherwise use target names. Duplicate gene names will be removed.
-#' @param split_assays If TRUE, split the transcript and protein barcoding targets into separate Seurat assays ("Transcript" and "Protein", respectively).
+#' @param split_assays If TRUE, split the transcript and protein barcoding targets into separate Seurat assays ("RNA" and "Protein", respectively).
 #' @return Seurat object
 #' @export
 cytoprofiling_to_seurat <- function(df, panel_json, use_gene_names = FALSE, split_assays = FALSE) {
@@ -31,18 +31,18 @@ cytoprofiling_to_seurat <- function(df, panel_json, use_gene_names = FALSE, spli
   transcript_counts <- t(df[, transcript_features])
   colnames(transcript_counts) <- df[, "Cell"]
 
+  if (use_gene_names) {
+    rownames(transcript_counts) <- sapply(rownames(transcript_counts), function(x) {
+      strsplit(x, ".", fixed = TRUE)[[1]][1]
+    }, USE.NAMES = FALSE)
+  }
+
   assay.name <- if (split_assays) "RNA" else "Barcoding"
 
   result <- Seurat::CreateSeuratObject(transcript_counts, assay = assay.name)
   result <- Seurat::AddMetaData(result, as.factor(df[, c("Well")]), col.name = "Well")
   result <- Seurat::AddMetaData(result, as.factor(df[, c("WellLabel")]), col.name = "WellLabel")
   Seurat::Idents(result) <- result$WellLabel
-
-  if (use_gene_names) {
-    rownames(result[[assay.name]]) <- sapply(rownames(result[[assay.name]]), function(x) {
-      strsplit(x, ".", fixed = TRUE)[[1]][1]
-    })
-  }
 
   if (split_assays) {
     protein_targets <- barcoding_targets[sapply(barcoding_targets, function(x) x$TargetType == "Protein")]
@@ -60,13 +60,13 @@ cytoprofiling_to_seurat <- function(df, panel_json, use_gene_names = FALSE, spli
     protein_counts <- t(df[, protein_features])
     colnames(protein_counts) <- df[, "Cell"]
 
-    result[["Protein"]] <- Seurat::CreateAssayObject(counts = protein_counts, assay = "Protein")
-
     if (use_gene_names) {
-      rownames(result[["Protein"]]) <- sapply(rownames(result[["Protein"]]), function(x) {
+      rownames(protein_counts) <- sapply(rownames(protein_counts), function(x) {
         strsplit(x, ".", fixed = TRUE)[[1]][1]
-      })
+      }, USE.NAMES = FALSE)
     }
+
+    result[["Protein"]] <- Seurat::CreateAssayObject(counts = protein_counts, assay = "Protein")
   }
 
   return(result)