+++ src/R/cytoprofiling/changelog.md
@@ -1,3 +1,12 @@
+# 1.2.2
+2026/04/14
+
+* Fix issue with export to Seurat using gene names with updated Seurat objects
+* Fix documentation of filter_cells function
+* Fix documentation of cytoprofiling_to_seurat function
+* Fix behavior of normalize_cells_by_aggregated_counts function when supplying list of normalization targets
+
+
 # 1.2.1
 2025/10/14